Add a openSUSE Leap 42.2 stemcell
The openSUSE stemcell is built in Docker container like the other
distributions.
A custom Docker image is needed because the other distributions
do not provide Kiwi which is needed for creating the openSUSE base
image.

For building details see:
ci/docker/suse-os-image-stemcell-builder/README.md

+++ stemcell_builder/stages/image_install_grub/apply.sh
@@ -79,9 +79,18 @@ if ! is_ppc64le; then
     # install bootsector into disk image file
     run_in_chroot ${image_mount_point} "grub2-install -v --no-floppy --grub-mkdevicemap=/device.map --target=i386-pc ${device}"
 
-    cat >${image_mount_point}/etc/default/grub <<EOF
+    # Enable password-less booting in openSUSE, only editing the boot menu needs to be restricted
+    if [ -f ${image_mount_point}/etc/SuSE-release ]; then
+      run_in_chroot ${image_mount_point} "sed -i 's/CLASS=\\\"--class gnu-linux --class gnu --class os\\\"/CLASS=\\\"--class gnu-linux --class gnu --class os --unrestricted\\\"/' /etc/grub.d/10_linux"
+
+      cat >${image_mount_point}/etc/default/grub <<EOF
+GRUB_CMDLINE_LINUX="vconsole.keymap=us net.ifnames=0 crashkernel=auto selinux=0 plymouth.enable=0 console=ttyS0,115200n8 earlyprintk=ttyS0 rootdelay=300 audit=1 cgroup_enable=memory swapaccount=1"
+EOF
+    else
+      cat >${image_mount_point}/etc/default/grub <<EOF
 GRUB_CMDLINE_LINUX="vconsole.keymap=us net.ifnames=0 crashkernel=auto selinux=0 plymouth.enable=0 console=ttyS0,115200n8 earlyprintk=ttyS0 rootdelay=300 audit=1"
 EOF
+    fi
 
     # we use a random password to prevent user from editing the boot menu
     pbkdf2_password=`run_in_chroot ${image_mount_point} "echo -e '${random_password}\n${random_password}' | grub2-mkpasswd-pbkdf2 | grep -Eo 'grub.pbkdf2.sha512.*'"`
@@ -176,6 +185,14 @@ then
 # /etc/fstab Created by BOSH Stemcell Builder
 UUID=${uuid} / ext4 defaults 1 1
 FSTAB
+elif [ -f ${image_mount_point}/etc/SuSE-release ] # openSUSE
+then
+  initrd_file="initramfs-${kernel_version}.img"
+  os_name=$(cat ${image_mount_point}/etc/SuSE-release)
+  cat > ${image_mount_point}/etc/fstab <<FSTAB
+# /etc/fstab Created by BOSH Stemcell Builder
+UUID=${uuid} / ext4 defaults 1 1
+FSTAB
 else
   echo "Unknown OS, exiting"
   exit 2
@@ -223,6 +240,18 @@ title ${os_name} (${kernel_version})
   kernel /boot/vmlinuz-${kernel_version} ro root=UUID=${uuid} net.ifnames=0 plymouth.enable=0 selinux=0 console=tty0 console=ttyS0,115200n8 earlyprintk=ttyS0 rootdelay=300 ipv6.disable=1 audit=1
   initrd /boot/${initrd_file}
 GRUB_CONF
+
+elif [ -f ${image_mount_point}/etc/SuSE-release ] # openSUSE
+then
+  cat > ${image_mount_point}/boot/grub/grub.conf <<GRUB_CONF
+default=0
+timeout=1
+title ${os_name} (${kernel_version})
+  root (hd0,0)
+  kernel /boot/vmlinuz-${kernel_version} ro root=UUID=${uuid} net.ifnames=0 plymouth.enable=0 selinux=0 console=tty0 console=ttyS0,115200n8 earlyprintk=ttyS0 rootdelay=300 ipv6.disable=1 audit=1
+  initrd /boot/${initrd_file}
+GRUB_CONF
+
 else
   echo "Unknown OS, exiting"
   exit 2
